Add required python version to docs and test deps.
When running `make test` for the first time on a machine not regularly used for
building and testing python code chances are high that pytest and pytest-cov
aren't installed. This add them as a dependency to requirements.txt.

Also adding information on minimum python version to run the code. With
python3.10 I ran into the error message

`TypeError: Parameters to generic types must be types. Got <module 'github3.issues.event'`

With python3.11 this error was gone.