Switch average computation to numpy
This commit switches code for computing averages to using numpy.

This is a first step towards supporting more than one statistic for
summarizing lists of times for state change of issues.

Averages are easy to compute but sensitive to outliers. In particular
when thinking of SLAs for times for first responses it is often more
helpful to look at median, quartiles or even 90th percentile.