First stab at figuring out the number of very active mentors (#206)
* First stab at figuring out the number of very active mentors in a project.

* Fix format and linter errors.

* Add flake to requirements.

* Adds call to mentor counter and output in markdown.

This adds the call to mentor counter and displays the results in markdown
including first tests for this functionality.

* Make mentor counting configurable.

This adds two configuration options: One to enable mentor counting and one for
configuring how many comments a user needs to leave in discussions, PRs. and
issues to be counted as an active mentor.

* Adds mentor counting to json output and adds missing config.

This adds mentor counting output to json format. In addition this change makes
max number of comments to evaluate configurable as well as the cutoff for
heavily involved mentors.

* Fix merge conflicts.

* Fix linting errors.

* fix: linting fixes

Signed-off-by: Zack Koppert <zkoppert@github.com>

* 8 is reasonable number of attrs

Signed-off-by: Zack Koppert <zkoppert@github.com>

* Update test_most_active_mentors.py

Co-authored-by: Jason Meridth <jmeridth@github.com>

* Update config.py

Co-authored-by: Jason Meridth <jmeridth@github.com>

* Update config.py

Remove merge residual

* Update requirements.txt

Remove lib only needed for testing.

* Update issue_metrics.py

Co-authored-by: Jason Meridth <jmeridth@github.com>

* Update config.py

* Update config.py

set type of `enable_mentor_count` to `bool`

* Update test_config.py

change tests to handle boolean change of enable_mentor_count

---------

Signed-off-by: Zack Koppert <zkoppert@github.com>
Co-authored-by: Drost-Fromm <isabel.drost-fromm@mac-85cab.lan>
Co-authored-by: Zack Koppert <zkoppert@github.com>
Co-authored-by: Jason Meridth <jmeridth@github.com>
Co-authored-by: Jason Meridth <jmeridth@gmail.com>